(test) e2e: extract per-command test files and add shared helpers
Extract the monolithic app.e2e.test.ts (1087 lines) into 9 per-command
E2E test files so each feature is independently testable. Add shared
E2E helpers to @lhremote/core/testing to eliminate duplication across
all 22 test files.

New test files: app-service, find-app, list-accounts, check-status,
campaign-repository, instance-lifecycle, query-profile,
scrape-messaging-history, check-replies.

Shared helpers: resolveAccountId (asserts accounts exist instead of
silently skipping), forceStopInstance, assertDefined, getE2EPersonId.

Also adds test:e2e:file script using --testPathPattern for file-level
filtering (avoids loading all files when running a single test).
