+++ packages/e2e/src/campaign-execution.e2e.test.ts
@@ -2,14 +2,11 @@
 // Copyright (C) 2026 Oleksii PELYKH
 
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
-import { describeE2E, launchApp, quitApp, retryAsync } from "@lhremote/core/testing";
+import { assertDefined, describeE2E, forceStopInstance, launchApp, quitApp, resolveAccountId, retryAsync } from "@lhremote/core/testing";
 import {
-  type Account,
   type AppService,
-  killInstanceProcesses,
   LauncherService,
   startInstanceWithRecovery,
-  waitForInstanceShutdown,
 } from "@lhremote/core";
 
 // CLI handlers
@@ -39,12 +36,6 @@ import {
 } from "@lhremote/mcp/tools";
 import { createMockServer } from "@lhremote/mcp/testing";
 
-/** Type-narrowing assertion — fails the test with `message` when `value` is nullish. */
-function assertDefined<T>(value: T, message: string): asserts value is NonNullable<T> {
-  expect(value, message).toBeDefined();
-  expect(value, message).not.toBeNull();
-}
-
 /**
  * Campaign config with two actions so campaign-move-next can advance
  * a person from the first action to the second.
@@ -61,31 +52,10 @@ actions:
 /** Test person LH ID — https://www.linkedin.com/in/ollybriz/ */
 const TEST_PERSON_ID = 10996;
 
-/**
- * Stop the instance gracefully, falling back to SIGKILL if that fails.
- */
-async function forceStopInstance(
-  launcher: LauncherService,
-  accountId: number | undefined,
-  launcherPort: number,
-): Promise<void> {
-  if (accountId === undefined) return;
-
-  try {
-    await launcher.stopInstance(accountId);
-    await waitForInstanceShutdown(launcherPort);
-    return;
-  } catch {
-    // Graceful stop failed — escalate to OS kill
-  }
-
-  await killInstanceProcesses(launcherPort);
-}
-
 describeE2E("Campaign execution and monitoring", () => {
   let app: AppService;
   let port: number;
-  let accountId: number | undefined;
+  let accountId: number;
 
   beforeAll(async () => {
     const launched = await launchApp();
@@ -93,30 +63,24 @@ describeE2E("Campaign execution and monitoring", () => {
     port = launched.port;
 
     // Start an account instance — required by execution operations
+    accountId = await resolveAccountId(port);
+
     const launcher = new LauncherService(port);
     await retryAsync(() => launcher.connect(), { retries: 3, delay: 1_000 });
-    const accounts = await launcher.listAccounts();
-
-    if (accounts.length > 0) {
-      accountId = (accounts[0] as Account).id;
-      await startInstanceWithRecovery(launcher, accountId, port);
-    }
-
+    await startInstanceWithRecovery(launcher, accountId, port);
     launcher.disconnect();
   }, 120_000);
 
   afterAll(async () => {
     // Stop the instance before quitting
-    if (accountId !== undefined) {
-      const launcher = new LauncherService(port);
-      try {
-        await launcher.connect();
-        await forceStopInstance(launcher, accountId, port);
-      } catch {
-        // Best-effort cleanup
-      } finally {
-        launcher.disconnect();
-      }
+    const launcher = new LauncherService(port);
+    try {
+      await launcher.connect();
+      await forceStopInstance(launcher, accountId, port);
+    } catch {
+      // Best-effort cleanup
+    } finally {
+      launcher.disconnect();
     }
     await quitApp(app);
   }, 60_000);
