+++ packages/e2e/src/get-feed.e2e.test.ts
@@ -2,16 +2,13 @@
 // Copyright (C) 2026 Oleksii PELYKH
 
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
-import { describeE2E, launchApp, quitApp, retryAsync } from "@lhremote/core/testing";
+import { describeE2E, forceStopInstance, launchApp, quitApp, resolveAccountId, retryAsync } from "@lhremote/core/testing";
 import {
-  type Account,
   type AppService,
   discoverInstancePort,
   discoverTargets,
-  killInstanceProcesses,
   LauncherService,
   startInstanceWithRecovery,
-  waitForInstanceShutdown,
 } from "@lhremote/core";
 import type { FeedPost, GetFeedOutput } from "@lhremote/core";
 
@@ -22,31 +19,10 @@ import { handleGetFeed } from "@lhremote/cli/handlers";
 import { registerGetFeed } from "@lhremote/mcp/tools";
 import { createMockServer } from "@lhremote/mcp/testing";
 
-/**
- * Stop the instance gracefully, falling back to SIGKILL if that fails.
- */
-async function forceStopInstance(
-  launcher: LauncherService,
-  accountId: number | undefined,
-  launcherPort: number,
-): Promise<void> {
-  if (accountId === undefined) return;
-
-  try {
-    await launcher.stopInstance(accountId);
-    await waitForInstanceShutdown(launcherPort);
-    return;
-  } catch {
-    // Graceful stop failed — escalate to OS kill
-  }
-
-  await killInstanceProcesses(launcherPort);
-}
-
 describeE2E("get-feed operation", () => {
   let app: AppService;
   let port: number;
-  let accountId: number | undefined;
+  let accountId: number;
   let cdpPort: number;
 
   beforeAll(async () => {
@@ -54,18 +30,12 @@ describeE2E("get-feed operation", () => {
     app = launched.app;
     port = launched.port;
 
+    accountId = await resolveAccountId(port);
+
     const launcher = new LauncherService(port);
     await retryAsync(() => launcher.connect(), { retries: 3, delay: 1_000 });
-    const accounts = await launcher.listAccounts();
-
-    if (accounts.length > 0) {
-      accountId = (accounts[0] as Account).id;
-      await startInstanceWithRecovery(launcher, accountId, port);
-    }
-
+    await startInstanceWithRecovery(launcher, accountId, port);
     launcher.disconnect();
-
-    if (accountId === undefined) return;
 
     // Discover the instance's dynamic CDP port
     const instancePort = await retryAsync(
@@ -94,16 +64,14 @@ describeE2E("get-feed operation", () => {
   }, 120_000);
 
   afterAll(async () => {
-    if (accountId !== undefined) {
-      const launcher = new LauncherService(port);
-      try {
-        await launcher.connect();
-        await forceStopInstance(launcher, accountId, port);
-      } catch {
-        // Best-effort cleanup
-      } finally {
-        launcher.disconnect();
-      }
+    const launcher = new LauncherService(port);
+    try {
+      await launcher.connect();
+      await forceStopInstance(launcher, accountId, port);
+    } catch {
+      // Best-effort cleanup
+    } finally {
+      launcher.disconnect();
     }
     await quitApp(app);
   }, 60_000);
